Mega refactory! Split function in module by cmd
Former-commit-id: f5f22a738597ca85b1757a40dec2743443cceba3
