+++ functions.py
@@ -21,13 +21,8 @@ from bs4 import BeautifulSoup
 import sqlite3
 import logging
 
-from pprint import pprint
-from sets import Set
-import calendar
-import locale
-import time
-
-locale.setlocale(locale.LC_ALL, 'it_IT.utf8')
+from module.lezioni import lezioni_cmd
+from module.esami import esami_cmd
 
 # Debug
 disable_chatid_logs = 1 #news, stats
@@ -45,118 +40,6 @@ tokenconf = open('config/token.conf', 'r').read()
 tokenconf = tokenconf.replace("\n", "")
 TOKEN = tokenconf      		#Token of your telegram bot that you created from @BotFather, write it on token.conf
 
-#Lezioni
-def output_lezioni(item):
-    daylist = list(calendar.day_name)
-
-    output = "*Insegnamento:* " + item["insegnamento"]
-    output += "\n*Aula:* " + item["aula"]
-
-    for day in daylist:
-        if(day.replace('ì','i') in item and item[day.replace('ì','i')] != ""):
-            output += "\n*" + day.title() + ":* " + item[day.replace('ì','i')]
-
-    output += "\n*Anno:* " + item["anno"] + "\n"
-
-    return output
-
-def condition_lezioni(items, condition, *arg):
-    output = Set()
-    for item in items:
-        if(arg):
-            if(arg[0] in item[condition].lower()):
-                output.add(output_lezioni(item))
-        else:
-            if(condition.replace('ì','i') in item and item[condition.replace('ì','i')] != ""):
-                output.add(output_lezioni(item))
-
-    return output
-
-def condition_mult_lezioni(items, days, years):
-    output = Set()
-    for item in items:
-        for day in days:
-            if( [year for year in years if year in item["anno"].lower()] ):
-                if(day.replace('ì','i') in item and item[day.replace('ì','i')] != ""):
-                    output.add(output_lezioni(item))
-    return output
-
-def lezioni_cmd(args, link):
-
-    output_str = "Poffarbacco, qualcosa non va. Segnalalo ai dev \contributors \n"
-
-    if(args):
-        output = Set()
-        r = requests.get(link)
-        if(r.status_code == requests.codes.ok):
-
-            items = r.json()["items"]
-            daylist = list(calendar.day_name)
-
-            #Clear arguments - Trasform all to lower case utf-8 (ì) - Remove word 'anno' and len<2
-            args = [x.lower().encode('utf-8') for x in args if len(x) > 2]
-            if 'anno' in args: args.remove('anno')
-
-            #Study case
-            if(len(args) == 1):
-
-                if(args[0] in daylist):
-                    output = condition_lezioni(items, args[0])
-
-                elif(args[0] == "oggi"):
-                    output = condition_lezioni(items, time.strftime("%A"))
-
-                elif(args[0] == "domani"):
-                    tomorrow_date = datetime.datetime.today() + datetime.timedelta(1)
-                    tomorrow_name = datetime.datetime.strftime(tomorrow_date,'%A')
-                    output = condition_lezioni(items, tomorrow_name)
-
-                elif(args[0] in ("primo", "secondo", "terzo")):
-                    output = condition_lezioni(items, "anno", args[0])
-
-                elif([item["insegnamento"].lower().find(args[0]) for item in items]):
-                    output = condition_lezioni(items, "insegnamento", args[0])
-
-                if(len(output)):
-                    output_str = '\n'.join(list(output))
-                    output_str += "\n_Risultati trovati: " + str(len(output)) + "/" + str(r.json()["status"]["length"]) + "_"
-                    output_str += "\n_Ultimo aggiornamento: " + r.json()["status"]["lastupdate"] + "_\n"
-                else:
-                    output_str = "Nessun risultato trovato :(\n"
-            elif(len(args) > 1):
-
-                #Create an array of days and years if in arguments
-                days = list(set(args).intersection(daylist))
-                years = list(set(args).intersection(("primo", "secondo", "terzo")))
-
-                if(days and years):
-                    output = condition_mult_lezioni(items, days, years)
-
-                elif("oggi" in args and years):
-                    day = [time.strftime("%A")]
-                    output = condition_mult_lezioni(items, day, years)
-
-                elif("domani" in args and years):
-                    tomorrow_date = datetime.datetime.today() + datetime.timedelta(1)
-                    tomorrow_name = datetime.datetime.strftime(tomorrow_date,'%A')
-                    day = [tomorrow_name]
-                    output = condition_mult_lezioni(items, day, years)
-
-                else:
-                    for arg in args:
-                        output = output.union(condition_lezioni(items, "insegnamento", arg))
-
-                if(len(output)):
-                    output_str = '\n'.join(list(output))
-                    output_str += "\n_Risultati trovati: " + str(len(output)) + "/" + str(r.json()["status"]["length"]) + "_"
-                    output_str += "\n_Ultimo aggiornamento: " + r.json()["status"]["lastupdate"] + "_\n"
-                else:
-                    output_str = "Nessun risultato trovato :(\n"
-    else:
-        output_str = "Inserisci almeno un parametro.\n"
-
-    return output_str
-
 def lezioni(bot, update, args, *m):
     checkLog(bot, update, "lezioni")
     if(m):
@@ -164,116 +47,6 @@ def lezioni(bot, update, args, *m):
     else:
         messageText = lezioni_cmd(args, 'http://188.213.170.165/PHP-DMI-API/result/lezioni_dmi.json')
     bot.sendMessage(chat_id=update.message.chat_id, text=messageText, parse_mode='Markdown')
-
-#Esami
-def output_esami(item, sessions):
-
-    output = "*Insegnamento:* " + item["insegnamento"]
-    output += "\n*Docenti:* " + item["docenti"]
-
-    for session in sessions:
-        appeal = [appeal for appeal in item[session] if appeal]
-        if(appeal):
-            output += "\n*" + session.title() + ":* " + " | ".join(appeal)
-
-    output += "\n*Anno:* " + item["anno"] + "\n"
-
-    return output
-
-def condition_esami(items, field, value, *session):
-    output = Set()
-    if(session):
-        for item in items:
-            if([appeal for appeal in item[value] if appeal]):
-                output.add(output_esami(item, [value]))
-    else:
-        for item in items:
-            if(value in item[field].lower()):
-                output.add(output_esami(item, ("prima", "seconda", "terza", "straordinaria")))
-
-    return output
-
-def esami_cmd(args, link):
-
-    output_str = "Poffarbacco, qualcosa non va. Segnalalo ai dev \contributors \n"
-
-    if(args):
-        output = Set()
-        r = requests.get(link)
-        if(r.status_code == requests.codes.ok):
-
-            items = r.json()["items"]
-            daylist = list(calendar.day_name)
-
-            #Clear arguments - Trasform all to lower case - Remove word 'anno', 'sessione'
-            args = [x.lower().encode('utf-8') for x in args if len(x) > 2]
-            if 'anno' in args: args.remove('anno')
-            if 'sessione' in args: args.remove('sessione')
-
-            #Study case
-            if(len(args) == 1):
-
-                if(args[0] in ("primo", "secondo", "terzo")):
-                    output = condition_esami(items, "anno", args[0])
-
-                elif(args[0] in ("prima", "seconda", "terza", "straordinaria")):
-                    output = condition_esami(items, "sessione", args[0], True)
-
-                elif([item["insegnamento"].lower().find(args[0]) for item in items]):
-                    output = condition_esami(items, "insegnamento", args[0])
-
-                if(len(output)):
-                    output_str = '\n'.join(list(output))
-                    output_str += "\n_Risultati trovati: " + str(len(output)) + "/" + str(r.json()["status"]["length"]) + "_"
-                    output_str += "\n_Ultimo aggiornamento: " + r.json()["status"]["lastupdate"] + "_\n"
-
-                else:
-                    output_str = "Nessun risultato trovato :(\n"
-
-            elif(len(args) > 1):
-
-                #Create an array of session and years if in arguments
-                sessions = list(set(args).intersection(("prima", "seconda", "terza", "straordinaria")))
-                years = list(set(args).intersection(("primo", "secondo", "terzo")))
-
-                if(sessions and years):
-                    for item in items:
-                        if(item["anno"].lower().replace("anno","").replace(" ", "") in years):
-                            if( [session for session in sessions if [appeal for appeal in item[session] if appeal]] ):
-                                output.add(output_esami(item, sessions))
-
-                elif(sessions and not years):
-                    #If years array is empty and session not, the other word are subjects
-                    subjects = [arg for arg in args if arg not in(sessions)]
-
-                    for item in items:
-                        if(subjects):
-                            for subject in subjects:
-                                if( [session for session in sessions if [appeal for appeal in item[session] if appeal]] ):
-                                    if(subject in item["insegnamento"].lower()):
-                                        output.add(output_esami(item, sessions))
-
-                        #List of session of all years [useless]
-                        '''
-                        else:
-                            if( [session for session in sessions if [appeal for appeal in item[session] if appeal]] ):
-                                output.add(output_esami(item, sessions))
-                        '''
-
-                else:
-                    for arg in args:
-                        output = output.union(condition_esami(items, "insegnamento", arg, False))
-
-                if(len(output)):
-                    output_str = '\n'.join(list(output))
-                    output_str += "\n_Risultati trovati: " + str(len(output)) + "/" + str(r.json()["status"]["length"]) + "_"
-                    output_str += "\n_Ultimo aggiornamento: " + r.json()["status"]["lastupdate"] + "_\n"
-                else:
-                    output_str = "Nessun risultato trovato :(\n"
-    else:
-        output_str = "Inserisci almeno un parametro.\n"
-
-    return output_str
 
 def esami(bot, update, args, *m):
     checkLog(bot, update, "esami")
